+++ packages/agtree/jest.config.ts
@@ -5,8 +5,6 @@ const config: Config = {
     testTimeout: 30000,
     testMatch: ['**/test/**/*.test.ts'],
     transform: {
-        // Enable importing YAML files while testing
-        '.yml': 'yaml-jest-transform',
         // Speed up tests by using SWC instead of Babel
         '.ts': '@swc/jest',
     },