+++ packages/agtree/test/validator/index.test.ts
@@ -6,12 +6,6 @@ import { VALIDATION_ERROR_PREFIX } from '../../src/validator/constants';
 import { AdblockSyntax } from '../../src/utils/adblockers';
 import { LIST_PARSE_ERROR_PREFIX } from '../../src/parser/misc/list-helpers';
 
-const DOCS_BASE_URL = {
-    ADG: 'https://adguard.app/kb/general/ad-filtering/create-own-filters/',
-    UBO: 'https://github.com/gorhill/uBlock/wiki/Static-filter-syntax',
-    ABP: 'https://help.adblockplus.org/',
-};
-
 /**
  * Returns modifier AST node for given rawModifier.
  *
@@ -1315,106 +1309,6 @@ describe('ModifierValidator', () => {
                 // third argument is 'false' for blocking rules
                 const validationResult = modifierValidator.validate(AdblockSyntax.Abp, modifier, false);
                 expect(validationResult.valid).toBeTruthy();
-            });
-        });
-    });
-
-    describe('getAdgDocumentationLink', () => {
-        describe('has docs', () => {
-            const modifiers = [
-                'denyallow',
-                'domain=example.com',
-                'third-party',
-                'important',
-                // deprecated
-                'empty',
-                'mp4',
-                // 'webrtc' is removed and not supported but it has docs url
-                'webrtc',
-            ];
-            test.each(modifiers)('%s', (rawModifier) => {
-                const modifier = getModifier(rawModifier);
-                const docsUrl = modifierValidator.getAdgDocumentationLink(modifier);
-                expect(docsUrl?.startsWith(DOCS_BASE_URL.ADG)).toBeTruthy();
-            });
-        });
-
-        describe('no docs', () => {
-            const modifiers = [
-                // not existent
-                'protobuf',
-                // not supported by ADG
-                'cname',
-            ];
-            test.each(modifiers)('%s', (rawModifier) => {
-                const modifier = getModifier(rawModifier);
-                const docsUrl = modifierValidator.getAdgDocumentationLink(modifier);
-                expect(docsUrl).toBeNull();
-            });
-        });
-    });
-
-    describe('getUboDocumentationLink', () => {
-        describe('has docs', () => {
-            const modifiers = [
-                'cname',
-                'from=example.com',
-                'third-party',
-                'important',
-                // deprecated modifiers
-                'empty',
-                'mp4',
-                // 'webrtc' is removed and not supported but it has docs url
-                'webrtc',
-            ];
-            test.each(modifiers)('%s', (rawModifier) => {
-                const modifier = getModifier(rawModifier);
-                const docsUrl = modifierValidator.getUboDocumentationLink(modifier);
-                expect(docsUrl?.startsWith(DOCS_BASE_URL.UBO)).toBeTruthy();
-            });
-        });
-
-        describe('no docs', () => {
-            const modifiers = [
-                // not existent
-                'protobuf',
-                // not supported by UBO
-                'removeheader',
-            ];
-            test.each(modifiers)('%s', (rawModifier) => {
-                const modifier = getModifier(rawModifier);
-                const docsUrl = modifierValidator.getUboDocumentationLink(modifier);
-                expect(docsUrl).toBeNull();
-            });
-        });
-    });
-
-    describe('getAbpDocumentationLink', () => {
-        describe('has docs', () => {
-            const modifiers = [
-                'domain=example.com',
-                'third-party',
-                'webrtc',
-            ];
-            test.each(modifiers)('%s', (rawModifier) => {
-                const modifier = getModifier(rawModifier);
-                const docsUrl = modifierValidator.getAbpDocumentationLink(modifier);
-                expect(docsUrl?.startsWith(DOCS_BASE_URL.ABP)).toBeTruthy();
-            });
-        });
-
-        describe('no docs', () => {
-            const modifiers = [
-                // not existent
-                'protobuf',
-                // not supported by ABP
-                'denyallow',
-                'important',
-            ];
-            test.each(modifiers)('%s', (rawModifier) => {
-                const modifier = getModifier(rawModifier);
-                const docsUrl = modifierValidator.getAbpDocumentationLink(modifier);
-                expect(docsUrl).toBeNull();
             });
         });
     });