+++ packages/tswebextension/src/lib/mv2/content-script/css-hits-counter.ts
@@ -7,7 +7,23 @@ import { type RuleInfo } from './rule-info';
 /**
  * Counted element data structure.
  */
-type ICountedElement = RuleInfo & { element: string };
+type ICountedElement = RuleInfo & { element: string | Element };
+
+const DEFAULT_ELEMENT_TO_STRING = true;
+
+/**
+ * CssHitsCounter options.
+ */
+interface CssHitsCounterOptions {
+    /**
+     * Flag determining if the element should be converted to a string.
+     * If true, the element is converted to a string.
+     * Otherwise, the element is left as is, which might be helpful in corelibs,
+     * where logs are printed in the developer tools console.
+     * By default, is true.
+     */
+    elementToString: boolean;
+}
 
 /**
  * Class represents collecting css style hits process.
@@ -60,12 +76,24 @@ export class CssHitsCounter {
     private countIsWorking = false;
 
     /**
+     * Flag determining if we should convert elements to string, or not.
+     * @private
+     */
+    private elementToString = DEFAULT_ELEMENT_TO_STRING;
+
+    /**
      * This function prepares calculation of css hits.
      * We are waiting for 'load' event and start calculation.
      *
      * @param callback Which receives {@link ICountedElement} and handles counted css hits.
+     * @param options CssHitsCounter options.
      */
-    constructor(callback: (x: ICountedElement[]) => void) {
+    constructor(callback: (x: ICountedElement[]) => void, options?: CssHitsCounterOptions) {
+        if (options) {
+            const { elementToString } = options;
+            this.elementToString = elementToString;
+        }
+
         this.onCssHitsFoundCallback = callback;
 
         if (document.readyState === 'complete'
@@ -108,10 +136,14 @@ export class CssHitsCounter {
 
                     const { filterId, ruleIndex } = ruleInfo;
                     if (filterId !== undefined && ruleIndex !== undefined) {
+                        const element = this.elementToString
+                            ? ElementUtils.elementToString(affectedEl.node)
+                            : affectedEl.node;
+
                         result.push({
                             filterId,
                             ruleIndex,
-                            element: ElementUtils.elementToString(affectedEl.node),
+                            element,
                         });
 
                         // clear style content to avoid duplicate counting
@@ -249,7 +281,7 @@ export class CssHitsCounter {
             result.push({
                 filterId,
                 ruleIndex,
-                element: ElementUtils.elementToString(element),
+                element: this.elementToString ? ElementUtils.elementToString(element) : element,
             });
         }